backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
rembg>=2.0.55
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
aiofiles==23.2.1
